+++ all-is-cubes-base/src/raycast/tests.rs
@@ -421,7 +421,7 @@ fn intersection_point_random_test() {
     // A one-cube box, so that all possible rays should either intersect
     // exactly this cube, or none at all.
     let bounds = GridAab::from_lower_size([0, 0, 0], [1, 1, 1]);
-    let ray_origins: Aab = bounds.expand(FaceMap::repeat(1)).to_free();
+    let ray_origins: Aab = bounds.expand(FaceMap::splat(1)).to_free();
 
     let mut rng = rand_xoshiro::Xoshiro256Plus::seed_from_u64(0);
     for _ in 0..1000 {